Moving Views.java to shared folder.
