+++ Paco-Server/src/com/google/sampling/experiential/server/CloudSqlSearchServlet.java
@@ -27,11 +27,11 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.sampling.experiential.datastore.EventServerColumns;
 import com.google.sampling.experiential.shared.EventDAO;
-import com.google.sampling.experiential.shared.Views;
 import com.pacoapp.paco.shared.model2.EventBaseColumns;
 import com.pacoapp.paco.shared.model2.JsonConverter;
 import com.pacoapp.paco.shared.model2.OutputBaseColumns;
 import com.pacoapp.paco.shared.model2.SQLQuery;
+import com.pacoapp.paco.shared.model2.Views;
 import com.pacoapp.paco.shared.util.Constants;
 import com.pacoapp.paco.shared.util.ErrorMessages;
 import com.pacoapp.paco.shared.util.QueryJsonParser;
