+++ Paco-Server/src/com/google/sampling/experiential/server/EventServlet.java
@@ -52,9 +52,9 @@ import com.google.common.collect.Maps;
 import com.google.sampling.experiential.model.Event;
 import com.google.sampling.experiential.model.PhotoBlob;
 import com.google.sampling.experiential.shared.EventDAO;
-import com.google.sampling.experiential.shared.Views;
 import com.google.sampling.experiential.shared.WhatDAO;
 import com.pacoapp.paco.shared.model2.JsonConverter;
+import com.pacoapp.paco.shared.model2.Views;
 
 /**
  * Servlet that answers queries for Events.
