+++ Paco-Server/src/com/google/sampling/experiential/server/JSONBlobWriter.java
@@ -32,9 +32,9 @@ import com.google.sampling.experiential.model.Event;
 import com.google.sampling.experiential.model.PhotoBlob;
 import com.google.sampling.experiential.shared.EventDAO;
 import com.google.sampling.experiential.shared.TimeUtil;
-import com.google.sampling.experiential.shared.Views;
 import com.google.sampling.experiential.shared.WhatDAO;
 import com.pacoapp.paco.shared.model2.JsonConverter;
+import com.pacoapp.paco.shared.model2.Views;
 
 public class JSONBlobWriter {
