+++ Shared/src/com/pacoapp/paco/shared/model2/Views.java
@@ -1,4 +1,4 @@
-package com.google.sampling.experiential.shared;
+package com.pacoapp.paco.shared.model2;
 
 public class Views {
   public static class V4 {
